+++ src/SBInclinedSersic.cpp
@@ -324,7 +324,8 @@ namespace galsim {
         return _flux * kValueHelper(kx,ky);
     }
 
-    void SBInclinedSersic::SBInclinedSersicImpl::fillKImage(ImageView<std::complex<double> > im,
+    template <typename T>
+    void SBInclinedSersic::SBInclinedSersicImpl::fillKImage(ImageView<std::complex<T> > im,
                                                 double kx0, double dkx, int izero,
                                                 double ky0, double dky, int jzero) const
     {
@@ -338,7 +339,7 @@ namespace galsim {
             xdbg<<"Non-Quadrant\n";
             const int m = im.getNCol();
             const int n = im.getNRow();
-            std::complex<double>* ptr = im.getData();
+            std::complex<T>* ptr = im.getData();
             int skip = im.getNSkip();
             assert(im.getStep() == 1);
 
@@ -355,7 +356,8 @@ namespace galsim {
         }
     }
 
-    void SBInclinedSersic::SBInclinedSersicImpl::fillKImage(ImageView<std::complex<double> > im,
+    template <typename T>
+    void SBInclinedSersic::SBInclinedSersicImpl::fillKImage(ImageView<std::complex<T> > im,
                                                 double kx0, double dkx, double dkxy,
                                                 double ky0, double dky, double dkyx) const
     {
@@ -364,7 +366,7 @@ namespace galsim {
         dbg<<"ky = "<<ky0<<" + i * "<<dkyx<<" + j * "<<dky<<std::endl;
         const int m = im.getNCol();
         const int n = im.getNRow();
-        std::complex<double>* ptr = im.getData();
+        std::complex<T>* ptr = im.getData();
         int skip = im.getNSkip();
         assert(im.getStep() == 1);